+++ services/121-service/src/programs/dto/create-program-question.dto.ts
@@ -29,6 +29,7 @@ export class CreateProgramQuestionDto {
     AnswerTypes.tel,
     AnswerTypes.text,
     AnswerTypes.date,
+    AnswerTypes.multiSelect,
   ])
   public readonly answerType: string;
   @ApiProperty()