+++ services/121-service/src/registration/dto/custom-data.dto.ts
@@ -1,6 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 
 import { Length, IsString, IsNotEmpty } from 'class-validator';
+import { IsRegistrationDataValidType } from '../validator/registration-data-type.validator';
 
 export class CustomDataDto {
   @ApiProperty({ example: '910c50be-f131-4b53-b06b-6506a40a2734' })
@@ -12,6 +13,6 @@ export class CustomDataDto {
   public readonly key: string;
   @ApiProperty({ example: '31600000000' })
   @IsNotEmpty()
-  @IsString()
-  public readonly value: string;
+  @IsRegistrationDataValidType({ referenceId: 'referenceId', attribute: 'key' })
+  public readonly value: string | string[];
 }