+++ services/121-service/src/registration/dto/store-program-answers.dto.ts
@@ -1,11 +1,13 @@
 import { ApiProperty } from '@nestjs/swagger';
+import { Type } from 'class-transformer';
 import {
   IsArray,
   IsNotEmpty,
-  IsNumber,
   IsString,
   Length,
+  ValidateNested,
 } from 'class-validator';
+import { IsRegistrationDataValidType } from '../validator/registration-data-type.validator';
 
 export class ProgramAnswer {
   @ApiProperty()
@@ -13,16 +15,30 @@ export class ProgramAnswer {
   @IsString()
   public readonly programQuestionName: string;
   @ApiProperty()
-  @IsNotEmpty()
-  @IsNumber()
-  public programAnswer: string;
+  @IsRegistrationDataValidType({
+    referenceId: 'referenceId',
+    attribute: 'programQuestionName',
+  })
+  public programAnswer: string | string[];
 }
 
 export class StoreProgramAnswersDto {
   @ApiProperty({ example: '910c50be-f131-4b53-b06b-6506a40a2734' })
   @Length(29, 36)
   public readonly referenceId: string;
-  @ApiProperty()
+  @ApiProperty({
+    example: {
+      referenceId: '910c50be-f131-4b53-b06b-6506a40a2734',
+      programAnswers: [
+        {
+          programQuestionName: 'name',
+          programAnswer: 'example',
+        },
+      ],
+    },
+  })
   @IsArray()
+  @ValidateNested({ each: true })
+  @Type(() => ProgramAnswer)
   public readonly programAnswers: ProgramAnswer[];
 }