+++ services/121-service/src/scripts/seed-program-test.ts
@@ -1,0 +1,44 @@
+import { Injectable } from '@nestjs/common';
+import { InterfaceScript } from './scripts.module';
+import { Connection } from 'typeorm';
+
+import { SeedHelper } from './seed-helper';
+import { SeedInit } from './seed-init';
+
+import fspBank from '../../seed-data/fsp/fsp-bank.json';
+import fspMixedAttributes from '../../seed-data/fsp/fsp-mixed-attributes.json';
+import fspNoAttributes from '../../seed-data/fsp/fsp-no-attributes.json';
+import fspIntersolve from '../../seed-data/fsp/fsp-intersolve.json';
+import fspAfricasTalking from '../../seed-data/fsp/fsp-africas-talking.json';
+
+import programTest from '../../seed-data/program/program-test.json';
+import instanceDemo from '../../seed-data/instance/instance-demo.json';
+
+@Injectable()
+export class SeedTestProgram implements InterfaceScript {
+  public constructor(private connection: Connection) {}
+
+  private readonly seedHelper = new SeedHelper(this.connection);
+
+  public async run(): Promise<void> {
+    const seedInit = await new SeedInit(this.connection);
+    await seedInit.run();
+
+    // ***** CREATE FINANCIAL SERVICE PROVIDERS *****
+    await this.seedHelper.addFsp(fspIntersolve);
+    await this.seedHelper.addFsp(fspAfricasTalking);
+    await this.seedHelper.addFsp(fspBank);
+    await this.seedHelper.addFsp(fspMixedAttributes);
+    await this.seedHelper.addFsp(fspNoAttributes);
+
+    // ***** CREATE PROGRAM *****
+    const program = await this.seedHelper.addProgram(programTest);
+
+    this.seedHelper.addDefaultUsers(program, true);
+
+    // ***** CREATE INSTANCE *****
+    await this.seedHelper.addInstance(instanceDemo);
+  }
+}
+
+export default SeedTestProgram;